requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
attrs==20.2.0
distlib==0.3.1
Django==3.1.2
django-colorfield==0.3.2
django-cors-headers==3.5.0
django-crispy-forms==1.10.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-pwa==1.0.10
djangorestframework==3.12.2
filelock==3.0.12
importlib-metadata==2.0.0
iniconfig==1.1.1
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
packaging==20.4
Pillow==8.0.0
pluggy==0.13.1
py==1.9.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.1.2
pytest-django==4.1.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
virtualenv==20.1.0
wrapt==1.12.1
zipp==3.4.0